src/main/java/com/tdd/katas/microservices/customerservice/service/CustomerServiceImpl.java
package com.tdd.katas.microservices.customerservice.service;

import com.tdd.katas.microservices.customerservice.model.CustomerData;
import com.tdd.katas.microservices.customerservice.repository.CustomerRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class CustomerServiceImpl implements CustomerService {

    @Autowired
    private CustomerRepository customerRepository;

    @Override
    public CustomerData getCustomerData(String customer_id) {
        return customerRepository.getCustomerData(customer_id);
    }
}
